+++ src/plugins/core/widgets/links/index.ts
@@ -1,0 +1,2 @@
+export { default as Links } from './Links';
+export { default as LinksSettings} from './LinksSettings';