+++ src/plugins/core/widgets/links/interfaces.ts
@@ -1,0 +1,7 @@
+export interface Link {
+  url: string;
+}
+
+export interface Settings {
+  links: Link[];
+}