+++ src/plugins/index.ts
@@ -2,7 +2,7 @@ export * from './interfaces';
 export * from './registry';
 
 // Import core plugins
-import './core';
+import './core/register';
 
 // Import extra plugins with external integrations
-import './extra';
+import './extra/register';